requirements.txt
torch>=2.1.0
torchvision>=0.16.0
einops>=0.7.0
pyyaml>=6.0
opencv-python>=4.8.0
numpy>=1.24.0
